requirements.txt
dogpile.cache
lxml
jsonpointer
matplotlib
requests

